must patch pointer type
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@34848 52c4871e-980c-0410-b1e0-e73912ce01f8

Former-commit-id: 9704e001b81e1704d714a44d8e9c26e66972f11d

+++ src/router/shared/bcmtimer.h
@@ -25,8 +25,8 @@
 /*
  * timer ID 
  */
-typedef unsigned int bcm_timer_module_id;
-typedef unsigned int bcm_timer_id;
+typedef unsigned long bcm_timer_module_id;
+typedef unsigned long bcm_timer_id;
 
 /*
  * timer callback
